💄 Fixed font style settings
