+++ src/components/utils/base/btns/ssRouterBtn.vue
@@ -1,42 +1,16 @@
 <script setup lang="ts">
-import { useQuasar } from 'quasar';
+import SsBtn from './ssBtn.vue';
 interface Prop {
   to: string;
 }
 defineProps<Prop>();
-
-const $q = useQuasar();
-const isDark = () => $q.dark.isActive;
 </script>
 
 <template>
-  <q-btn
-    outline
-    :to="to"
-    no-caps
-    :color="isDark() ? 'primary' : ''"
-    :class="isDark() ? '' : 'light'"
-    class="btn"
-  >
-    <div class="row items-top text">
+  <SsBtn outline :to="to" no-caps>
+    <div class="row items-top">
       <q-icon name="navigate_next" />
       <slot />
     </div>
-  </q-btn>
+  </SsBtn>
 </template>
-
-<style scoped lang="scss">
-.light {
-  background-color: $primary !important;
-}
-
-.text {
-  font-family: 'Genokaku JP';
-  font-weight: bold;
-  font-size: 0.9rem;
-}
-
-.btn {
-  border-radius: 5px;
-}
-</style>
