+++ src/css/app.scss
@@ -16,7 +16,7 @@
   src: url(./fonts/SourceHanSans-VF.otf.woff2);
 }
 
-* {
+body {
   font-family: 'Mamelon';
   font-size: 18px;
 }
